Main Menu Update
Added a Main Menu Scene with two canvases, one for the Main Menu and one for an Options Menu.

Added an audio player for eventual music that can carry between scenes, as well as a volume mixer and separate volume sliders for Master Volume, Music Volume and SFX Volume. May want to add Voice as a separate volume at some point.

REMEMBER to add all the scenes in the Build Settings menu in unity so that the menu script can actually access them for switching, currently the start button takes to my test scene just to see if it was working.